Improved Arrow Controls
the starting position, i.e the position that serves as the center of the
bow for the arrow, is a little bit above the center of the action area
(previously it was set to where ever the player left the arrow
stationary).  This position is set once the arrow is in the action area.